Initial commit.
- Add model.
- Add some Ui for debugging.

The model can recognize most of the characters on the image expect for 1 because there is no horizontal
padding in the image which is apparently required by the model. The UI will be removed once everything
is tested and is working. The captcha data is hardcoded for now. In the future ContentProvider (most likely)
will be used for passing/retrieving data to/from the app.